Remove old testing class
